% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{GetData}
\alias{GetData}
\title{Get Marquette BB Data}
\usage{
GetData(season)
}
\arguments{
\item{season}{String indicating which season you want to pull data for. 
e.g. set to "2017" for 2016-17 season stats}
}
\description{
This function scrapes ESPN.com and returns some player statistics for
             Marquette's men's basketball team.
}
