% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{NamesToAbbreviations}
\alias{NamesToAbbreviations}
\title{Names to abbreviations}
\usage{
NamesToAbbreviations(nameVec)
}
\arguments{
\item{nameVec}{A character vector of player names}
}
\description{
This function takes a vector of names and returns a vector of corresponding initials.
}
